RR-2785 - Improved PrisonerReceivedEventDueToTransferTest
